Image orientation moves to vips.
Tests for vips based orientation.
