+++ common-lib/src/main/scala/com/gu/mediaservice/lib/imaging/ImageOperations.scala
@@ -15,7 +15,7 @@ import com.gu.mediaservice.model._
 
 import scala.concurrent.{ExecutionContext, Future}
 import scala.sys.process._
-import app.photofox.vipsffm.{VImage, Vips, VipsOption}
+import app.photofox.vipsffm.{VImage, Vips, VipsHelper, VipsOption}
 
 
 case class ExportResult(id: String, masterCrop: Asset, othersizings: List[Asset])
@@ -347,6 +347,20 @@ object ImageOperations extends GridLogging {
         dimensions = Some(Dimensions(width = width, height = height))
       }
       dimensions
+    }
+  }
+
+  def orientation(sourceFile: File)(implicit ec: ExecutionContext): Future[Option[OrientationMetadata]] = {
+    Future {
+      var orientation: Option[OrientationMetadata] = None
+      Vips.run { arena =>
+        val image = VImage.newFromFile(arena, sourceFile.getAbsolutePath)
+        val exifOrientation = VipsHelper.image_get_orientation(image.getUnsafeStructAddress)
+        orientation = Some(OrientationMetadata(
+          exifOrientation = Some(exifOrientation)
+        ))
+      }
+      Seq(orientation).flatten.find(_.transformsImage())
     }
   }
