+++ image-loader/app/lib/imaging/FileMetadataReader.scala
@@ -183,23 +183,6 @@ object FileMetadataReader extends GridLogging {
 
   private def dateToUTCString(date: DateTime): String = ISODateTimeFormat.dateTime.print(date.withZone(DateTimeZone.UTC))
 
-
-  def orientation(image: File)(implicit logMarker: LogMarker): Future[Option[OrientationMetadata]] = {
-    for {
-      metadata <- readMetadata(image)
-    } yield {
-
-      for {
-        exifDirectory <- Option(metadata.getFirstDirectoryOfType(classOf[ExifIFD0Directory]))
-        exifOrientation <- Option(exifDirectory.getInteger(ExifDirectoryBase.TAG_ORIENTATION))
-        orientation = OrientationMetadata(exifOrientation = Some(exifOrientation))
-        orientationWhichTransformsImage <- Seq(orientation).find(_.transformsImage())
-      } yield {
-        orientationWhichTransformsImage
-      }
-    }
-  }
-
   def getColorModelInformation(image: File, metadata: Metadata, mimeType: MimeType)(implicit logMarker: LogMarker): Future[Map[String, String]] = {
     val stopWatch = Stopwatch.start
     val source = addImage(image)
