+++ image-loader/app/model/Uploader.scala
@@ -139,8 +139,6 @@ object Uploader extends GridLogging {
 
     val tempDirForRequest: File = Files.createTempDirectory(deps.config.tempDir.toPath, "upload").toFile
 
-    val sourceOrientationMetadataFuture = FileMetadataReader.orientation(uploadRequest.tempFile)
-
     val storableOriginalImage = StorableOriginalImage(
       uploadRequest.imageId,
       uploadRequest.tempFile,
@@ -157,7 +155,7 @@ object Uploader extends GridLogging {
       s3Source <- sourceStoreFuture
       mergedUploadRequest = patchUploadRequestWithS3Metadata(uploadRequest, s3Source)
       sourceDimensions <- ImageOperations.dimensions(uploadRequest.tempFile)
-      sourceOrientationMetadata <- sourceOrientationMetadataFuture
+      sourceOrientationMetadata <- ImageOperations.orientation(uploadRequest.tempFile)
       thumbViewableImage <- createThumbFuture(browserViewableImage, deps, tempDirForRequest, uploadRequest.instance, orientationMetadata = sourceOrientationMetadata)
       s3Thumb <- storeOrProjectThumbFile(thumbViewableImage)
       maybeStorableOptimisedImage <- getStorableOptimisedImage(
